+++ packages/matter-node.js/test/storage/StorageBackendDiskTest.ts
@@ -20,25 +20,34 @@ describe("StorageInMemory", () => {
     it("write and read success", () => {
         const storage = new StorageBackendDisk(TEST_STORAGE_LOCATION);
 
-        storage.set("context", "key", "value");
+        storage.set(["context"], "key", "value");
 
-        const value = storage.get("context", "key");
+        const value = storage.get(["context"], "key");
+        assert.equal(value, "value");
+    });
+
+    it("write and read success with multiple context levels", () => {
+        const storage = new StorageBackendDisk(TEST_STORAGE_LOCATION);
+
+        storage.set(["context", "subcontext", "subsubcontext"], "key", "value");
+
+        const value = storage.get(["context", "subcontext", "subsubcontext"], "key");
         assert.equal(value, "value");
     });
 
     it("Throws error when context is empty on set", () => {
         assert.throws(() => {
             const storage = new StorageBackendDisk(TEST_STORAGE_LOCATION);
-            storage.set("", "key", "value");
+            storage.set([""], "key", "value");
         }, {
-            message: "Context and key must not be empty strings!"
+            message: "Context must not be an empty string!"
         });
     });
 
     it("Throws error when key is empty on set", () => {
         assert.throws(() => {
             const storage = new StorageBackendDisk(TEST_STORAGE_LOCATION);
-            storage.set("context", "", "value");
+            storage.set(["context"], "", "value");
         }, {
             message: "Context and key must not be empty strings!"
         });
@@ -47,16 +56,25 @@ describe("StorageInMemory", () => {
     it("Throws error when context is empty on get", () => {
         assert.throws(() => {
             const storage = new StorageBackendDisk(TEST_STORAGE_LOCATION);
-            storage.get("", "key");
+            storage.get([""], "key");
         }, {
-            message: "Context and key must not be empty strings!"
+            message: "Context must not be an empty string!"
+        });
+    });
+
+    it("Throws error when context is empty on get with subcontext", () => {
+        assert.throws(() => {
+            const storage = new StorageBackendDisk(TEST_STORAGE_LOCATION);
+            storage.get(["ok", ""], "key");
+        }, {
+            message: "Context must not be an empty string!"
         });
     });
 
     it("Throws error when key is empty on get", () => {
         assert.throws(() => {
             const storage = new StorageBackendDisk(TEST_STORAGE_LOCATION);
-            storage.get("context", "");
+            storage.get(["context"], "");
         }, {
             message: "Context and key must not be empty strings!"
         });